+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,6 +1,6 @@
 ## What
 
-[Link to story](https://dsdmoj.atlassian.net/browse/LASB-XXX)
+[Link to story](https://dsdmoj.atlassian.net/browse/DSTEW-XXX)
 
 Describe what you did and why.
