+++ .github/workflows/build-test-pr.yml
@@ -44,7 +44,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: checkstyle-report
-          path: laa-data-stewardship-payments-service/build/reports/checkstyle
+          path: claims-data-service/build/reports/checkstyle
           retention-days: 14
 
       - name: Upload test report
@@ -52,7 +52,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-report
-          path: laa-data-stewardship-payments-service/build/reports/tests
+          path: claims-data-service/build/reports/tests
           retention-days: 14
 
       - name: Upload jacoco coverage report
@@ -60,7 +60,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: jacoco-coverage-report
-          path: laa-data-stewardship-payments-service/build/reports/jacoco
+          path: claims-data-service/build/reports/jacoco
           retention-days: 14
 
   vulnerability-scan-app:
