+++ .github/workflows/pr-merge-main.yml
@@ -38,7 +38,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-results
-          path: laa-data-stewardship-payments-service/build/test-results
+          path: claims-data-service/build/test-results
           retention-days: 14
 
       - name: Upload checkstyle report
@@ -46,7 +46,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: checkstyle-report
-          path: laa-data-stewardship-payments-service/build/reports/checkstyle
+          path: claims-data-service/build/reports/checkstyle
           retention-days: 14
 
       - name: Upload test report
@@ -54,7 +54,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: test-report
-          path: laa-data-stewardship-payments-service/build/reports/tests
+          path: claims-data-service/build/reports/tests
           retention-days: 14
 
       - name: Upload jacoco coverage report
@@ -62,7 +62,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: jacoco-coverage-report
-          path: laa-data-stewardship-payments-service/build/reports/jacoco
+          path: claims-data-service/build/reports/jacoco
           retention-days: 14
 
   vulnerability-report:
