+++ claims-data/docker-compose.yml
@@ -1,0 +1,6 @@
+services:
+  app:
+    build:
+      context: ..
+    ports:
+      - "8081:8080"
