+++ claims-data/service/src/main/resources/schema.sql
