--- docker-compose.yml
@@ -1,6 +1,0 @@
-services:
-  app:
-    build:
-      context: .
-    ports:
-      - "8080:8080"
