--- laa-data-stewardship-payments-service/src/main/java/uk/gov/justice/laa/datastewardship/payments/entity/ItemEntity.java
@@ -1,30 +1,0 @@
-package uk.gov.justice.laa.datastewardship.payments.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * The entity class for items.
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "ITEMS")
-public class ItemEntity {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  private String name;
-
-  private String description;
-}
