--- laa-data-stewardship-payments-service/src/main/resources/application.yml
@@ -1,44 +1,0 @@
-spring:
-  application:
-    name: LAA Data Stewardship Payments Application
-
-  # example database
-  datasource:
-    url: jdbc:h2:mem:itemsDb
-    driver-class-name: org.h2.Driver
-    username: sa
-    password: password
-  h2:
-    console:
-      enabled: true
-
-  #do not auto create db schema, use .sql scripts
-  jpa:
-    database-platform: org.hibernate.dialect.PostgreSQLDialect
-    show-sql: false
-    generate-ddl: false
-    hibernate:
-      ddl-auto: none
-
-# used by actuator info endpoint
-info:
-  app:
-    name: LAA Data Stewardship Payments Application
-    description: LAA Data Stewardship Payments Application that provides APIs to access all payments related data.
-    version: 1.0.0
-
-management:
-  # specify actuator endpoints to enable
-  endpoints:
-    web:
-      exposure:
-        include: health,info
-
-  # show detailed health status
-  endpoint:
-    health:
-      show-details: always
-
-  info:
-    env:
-      enabled: true
